refactor: Remoção de State desnecessário
